API reference

List of public functions exposed by CatanEngine

Catan.PlayerPublicView — Type

PlayerPublicView

Combines the public fields of Player, with some additional fields representing the info available publicly about the private fields. E.g. everyone knows how many dev cards each player has, but not which ones.

Catan.choose_building_location — Method

`choose_building_location(board::Board, players::Vector{PlayerPublicView}, player::RobotPlayer, candidates::Vector{Tuple{Int, Int}}, building_type::Symbol)::Tuple{Int,Int}`

candidates is guaranteed to be non-empty. This method is only called if there is a legal placement available.

Catan.generate_random_map — Method

Generate a random board conforming to the game constraints set in constants.jl. Save the generated map.

Catan.get_player_config — Function

```get_player_config(configs, team, key)```

Retrieves the key, potentially overridden for this team.

Catan.initialize_player — Method

initialize_player(board::Board, player::PlayerType)

This function is useful to do any one-time computations of the player as soon as the board is generated.

Catan.propose_trade_goods — Method

`propose_trade_goods(board::Board, players::Vector{PlayerType}, from_player::PlayerType, from_goods, to_goods)`

Handles the action of from_player offering from_goods in exchange for to_goods. Each of the players in players need to evaluate the proposition and decide if they will accept via choose_accept_trade.

Then, finally, from_player will decide among the accepted players, accepted_public, with whome he will carry out the trade.

Catan.read_map — Method

`read_map(configs::Dict)::Board`

Handles three cases: LOADMAP from file if specified, write the map to SAVEMAP if specified, and if neither are set, then we just generate the map and keep in memory.

Catan.run — Method

@docs run(configs::Dict)

Run one game, returning the board and the winner, accepting a config TOML file to overwrite any default configuration keys set in ./DefaultConfiguration.toml.

Catan.save_parameters_after_game_end — Method

save_parameters_after_game_end(board::Board, player::PlayerType)

After the game, store or update parameters based on the end state

Catan.setup_and_do_robot_game — Method

setup_and_do_robot_game(players::Vector{PlayerType}, savefile::Union{Nothing, String} = nothing)

If no savefile is passed, we use the standard format "testrobotgamesavefileyyyyMMdd_HHmmss.txt". If a savefile is passed, we use it to save the game state. If the file is nonempty, the game will replay up until the end of the save file and then continue to write ongoing game states to the file.

Catan.trade_goods_from_player — Method

`trade_goods_from_player(from_player::PlayerPublicView, from_goods::Vector{Symbol}, to_goods::Vector{Symbol})`

This implementation is useful in when a player is evaluating whether they should accept a trade. To see the